Register "Timer Record..."; break TransportMenus -> TimerRecordDialog
